emails_godaddy/requirements.txt
Flask==2.0.0
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.25
python-dotenv==0.17.1
